requirements.txt
adium-theme-ubuntu==0.3.4
amqp==2.4.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
cliapp==1.20160109
cmdtest==0.22
cryptography==1.2.3
decorator==4.3.2
defer==1.0.6
django-cors-headers==2.4.0
django-ipware==2.1.0
enum34==1.1.6
fluendo-eula==0.6
gWakeOnLan==0.5.1
gyp==0.1
html5lib==0.999
idna==2.8
influxdb==5.2.1
iniparse==0.4
ipaddress==1.0.16
ipdb==0.11
ipython==5.8.0
ipython-genutils==0.2.0
kombu==4.2.2.post1
lxml==3.5.0
Markdown==2.6.6
ndg-httpsclient==0.4.0
networkx==2.2
numpy==1.16.2
opencv-python==4.0.0.21
pandas==0.24.1
pathlib2==2.3.3
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyasn1==0.1.9
pycrypto==2.6.1
pycurl==7.43.0
Pygments==2.3.1
pygobject==3.20.0
pyliblzma==0.5.3
pyOpenSSL==0.15.1
pysqlite==1.0.1
python-apt==1.1.0b1+ubuntu0.16.4.2
python-dateutil==2.8.0
python-debian==0.1.27
python-distutils-extra==2.39
pytracking==0.2.0
pytz==2018.9
PyYAML==3.11
redis==3.1.0
requests==2.21.0
rpm-python==4.12.0.1
scandir==1.9.0
simplegeneric==0.8.1
six==1.12.0
traitlets==4.3.2
ttystatus==0.32
unity-lens-photos==1.0
urlgrabber==3.9.1
urllib3==1.24.1
vine==1.2.0
wcwidth==0.1.7
yum-metadata-parser==1.1.4
